CDumbPresentationOracle
